Move eth.LogFilter to own file.
